1. Changed to type long int:   N and its relatives (neq, Nlocal, Nglobal, etc.);   bandwidth parameters (ml, mu, smu, etc.);   internal indices (i, j, ngroups, etc.) that can have size ~ N;   pivot array, renamed lpivots;   last return flag optional output lsflag; with the following exceptions, where these are of type int:   N, ml, mu, etc. in the C user's call to CVLapackDense and CVLapackBand;   N etc. in the Fortran user's call to FCVLAPACKDENSE and FCVLAPACKBAND   (to be declared as Fortran equivalent of int). (But N, ml, mu arguments in LAPACK user Jacobian fn are still long int.) 2. Interface functions to Blas/Lapack routines declare local int copies of N etc. (intn etc.), and &intn is passed to the BLAS and LAPACK functions in place of &n, etc. 3. Direct linear solver memory block has arrays pivots (int) and lpivots (long int).  File sundials_direct.c has functions NewIntArray and NewLintArray, to create pivots and lpivots, respectively.  LAPACK interfaces use pivots, all other direct solvers use lpivots.

+++ include/cvode/cvode.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.12 $
- * $Date: 2007-11-26 16:19:58 $
+ * $Revision: 1.13 $
+ * $Date: 2010-12-01 22:10:38 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Scott D. Cohen, Alan C. Hindmarsh, Radu Serban
  *                and Dan Shumaker @ LLNL
@@ -769,7 +769,7 @@ SUNDIALS_EXPORT int CVodeGetNonlinSolvStats(void *cvode_mem, long int *nniters,
  * -----------------------------------------------------------------
  */
 
-SUNDIALS_EXPORT char *CVodeGetReturnFlagName(int flag);
+SUNDIALS_EXPORT char *CVodeGetReturnFlagName(long int flag);
 
 /*
  * -----------------------------------------------------------------
