+++ include/cvode/cvode_direct.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.5 $
- * $Date: 2008-04-18 19:42:36 $
+ * $Revision: 1.6 $
+ * $Date: 2010-12-01 22:10:38 $
  * ----------------------------------------------------------------- 
  * Programmer: Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -119,7 +119,7 @@ extern "C" {
  */
   
   
-typedef int (*CVDlsDenseJacFn)(int N, realtype t,
+typedef int (*CVDlsDenseJacFn)(long int N, realtype t,
 			       N_Vector y, N_Vector fy, 
 			       DlsMat Jac, void *user_data,
 			       N_Vector tmp1, N_Vector tmp2, N_Vector tmp3);
@@ -206,7 +206,7 @@ typedef int (*CVDlsDenseJacFn)(int N, realtype t,
  * -----------------------------------------------------------------
  */
 
-typedef int (*CVDlsBandJacFn)(int N, int mupper, int mlower,
+typedef int (*CVDlsBandJacFn)(long int N, long int mupper, long int mlower,
 			      realtype t, N_Vector y, N_Vector fy, 
 			      DlsMat Jac, void *user_data,
 			      N_Vector tmp1, N_Vector tmp2, N_Vector tmp3);
@@ -266,7 +266,7 @@ SUNDIALS_EXPORT int CVDlsSetBandJacFn(void *cvode_mem, CVDlsBandJacFn jac);
 SUNDIALS_EXPORT int CVDlsGetWorkSpace(void *cvode_mem, long int *lenrwLS, long int *leniwLS);
 SUNDIALS_EXPORT int CVDlsGetNumJacEvals(void *cvode_mem, long int *njevals);
 SUNDIALS_EXPORT int CVDlsGetNumRhsEvals(void *cvode_mem, long int *nfevalsLS);
-SUNDIALS_EXPORT int CVDlsGetLastFlag(void *cvode_mem, int *flag);
+SUNDIALS_EXPORT int CVDlsGetLastFlag(void *cvode_mem, long int *flag);
 
 /*
  * -----------------------------------------------------------------
@@ -275,7 +275,7 @@ SUNDIALS_EXPORT int CVDlsGetLastFlag(void *cvode_mem, int *flag);
  * -----------------------------------------------------------------
  */
 
-SUNDIALS_EXPORT char *CVDlsGetReturnFlagName(int flag);
+SUNDIALS_EXPORT char *CVDlsGetReturnFlagName(long int flag);
 
 
 #ifdef __cplusplus
