+++ include/ida/ida_dense.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.5 $
- * $Date: 2008-04-18 19:42:37 $
+ * $Revision: 1.6 $
+ * $Date: 2010-12-01 22:14:09 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Alan C. Hindmarsh and Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -49,7 +49,7 @@ extern "C" {
  * -----------------------------------------------------------------
  */
 
-SUNDIALS_EXPORT int IDADense(void *ida_mem, int Neq); 
+SUNDIALS_EXPORT int IDADense(void *ida_mem, long int Neq); 
 
 #ifdef __cplusplus
 }
